Extract ElementProc concern
